refactor: extract ResultTable DetailsTable
