+++ package.json
@@ -12,7 +12,7 @@
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "@types/jest": "^24.9.1",
-    "@types/node": "^12.12.57",
+    "@types/node": "^12.12.62",
     "@types/react-dom": "^16.9.8",
     "@types/react-router-dom": "^5.1.5",
     "bootstrap": "^4.5.2",
@@ -23,7 +23,7 @@
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
-    "typescript": "^3.7.5"
+    "typescript": "^3.9.7"
   },
   "scripts": {
     "start": "react-scripts start",
